Add Extensions to Gloo Validation: Rate Limit Configs (#7385)
* initial commit

* added in helm changes

* fix for helm tests

* tabs instead of spaces

* added test to verify that rate limit will error if written witih Validation on

* lets see those statuses

* using Sync() instead of Translate().  Make interface for TranlationSyncer one method.
Isolated SyncerTranslators to it's own validation service.

* fix issue with error of non-supportive resources. Also now rate limit applied test

* updated the testing, added flag for extension resources.

* stop focus on gateway tests

* get rid of supported concept and flags

* pass CI helm test

* check list GVK before we check gvks

* added in changelog

* get rid of extension validation flag

* pass server_test

* move test

* added in some tests for syncer validation tests

* update with review changes.

* unmarshalling the data will not include the namespaces status

* add ext-auth to validation

* get rid of TODO comment

* deprecated: IgnoreProxyValidationFailure

* update changelog

* get rid of duplicate error. So that warnings from the extension are visible.

* pass CI, changed will error when missing Ext-Auth to not error.

* update testing for ext-auth error

* update the tests to accept the VS that is missing ext-auth

* pass ext-auth e2e test

* get rid of focus

* update changelog with the IgnoreProxyValidationFailure

* Adding changelog file to new location

* Deleting changelog file from old location

* spelling on changelog

* adding in some comments

* updates from review, added comments and changelog changes

* test correction

* added test for Passing and Warning extentions

* Revert "pass ext-auth e2e test"

This reverts commit 53840e865f870d181076a8520948a782e2875194.

* Revert "update the tests to accept the VS that is missing ext-auth"

This reverts commit 24cf81d239cc85bf254012c0d2c11bb5cf74b5c3.

* Revert "update testing for ext-auth error"

This reverts commit 3faf329c904a5891170133a3940232a7009470a4.

* Revert "pass CI, changed will error when missing Ext-Auth to not error."

This reverts commit fb9b69c725e1395332580a82573951c9876294e3.

* Revert "get rid of duplicate error. So that warnings from the extension are visible."

This reverts commit fe1599609430c0ef02c2d3a588bfb9621ad31237.

* update changelog and change tests to support expected behavior

* update to show logs in the api emitter,  to get visuals on the proxy

* redact authConfigs from validation

* clean up some loose variables

* get rid of ext-auth test

* more clean up

* get rid of logging

* update the changelog

* get rid of ext-auth test

+++ projects/gateway/pkg/services/k8sadmission/validating_admission_webhook.go
@@ -20,8 +20,6 @@ import (
 	gloov1 "github.com/solo-io/gloo/projects/gloo/pkg/api/v1"
 	"github.com/solo-io/gloo/projects/gloo/pkg/api/v1/gloosnapshot"
 
-	"github.com/solo-io/solo-kit/pkg/utils/protoutils"
-
 	"github.com/solo-io/gloo/pkg/utils"
 
 	"go.opencensus.io/stats"
@@ -397,9 +395,16 @@ func (wh *gatewayValidationWebhook) validateAdmissionRequest(
 	if gvk == ListGVK {
 		return wh.validateList(ctx, admissionRequest.Object.Raw, dryRun)
 	}
+
+	if _, hit := gloosnapshot.ApiGvkToHashableResource[gvk]; !hit {
+		contextutils.LoggerFrom(ctx).Debugf("unsupported validation for resource namespace [%s] name [%s] group [%s] kind [%s]", ref.GetNamespace(), ref.GetName(), gvk.Group, gvk.Kind)
+		return &validation.Reports{}, nil
+	}
+
 	if isDelete {
 		return wh.deleteRef(ctx, gvk, ref, admissionRequest)
 	}
+
 	return wh.validateGvk(ctx, gvk, ref, admissionRequest)
 }
 
@@ -420,11 +425,6 @@ func (wh *gatewayValidationWebhook) deleteRef(ctx context.Context, gvk schema.Gr
 
 func (wh *gatewayValidationWebhook) validateGvk(ctx context.Context, gvk schema.GroupVersionKind, ref *core.ResourceRef, admissionRequest *v1beta1.AdmissionRequest) (*validation.Reports, *multierror.Error) {
 	var reports *validation.Reports
-	if !wh.validator.ModificationIsSupported(gvk) {
-		contextutils.LoggerFrom(ctx).Debugf("unsupported validation for resource namespace [%s] name [%s] group [%s] kind [%s]", ref.GetNamespace(), ref.GetName(), gvk.Group, gvk.Kind)
-		return &validation.Reports{}, nil
-	}
-
 	newResourceFunc := gloosnapshot.ApiGvkToHashableResource[gvk]
 
 	newResource := newResourceFunc()
@@ -464,7 +464,7 @@ func (wh *gatewayValidationWebhook) validateList(ctx context.Context, rawJson []
 func (wh *gatewayValidationWebhook) shouldValidateResource(ctx context.Context, admissionRequest *v1beta1.AdmissionRequest, resource, oldResource resources.HashableResource) (bool, error) {
 	logger := contextutils.LoggerFrom(ctx)
 
-	if err := protoutils.UnmarshalResource(admissionRequest.Object.Raw, resource); err != nil {
+	if err := validation.UnmarshalResource(admissionRequest.Object.Raw, resource); err != nil {
 		return false, &multierror.Error{Errors: []error{WrappedUnmarshalErr(err)}}
 	}
 	if skipValidationCheck(resource.GetMetadata().GetAnnotations()) {
@@ -478,7 +478,7 @@ func (wh *gatewayValidationWebhook) shouldValidateResource(ctx context.Context, 
 
 	// For update requests, we check to see if this is a status update
 	// If it is, we do not need to validate the resource
-	if err := protoutils.UnmarshalResource(admissionRequest.OldObject.Raw, oldResource); err != nil {
+	if err := validation.UnmarshalResource(admissionRequest.OldObject.Raw, oldResource); err != nil {
 		return false, &multierror.Error{Errors: []error{WrappedUnmarshalErr(err)}}
 	}
